refactor: use a main job to handle booking creation
