+++ app/jobs/bookings/create_first_checkin_mission_job.rb
@@ -1,0 +1,9 @@
+module Bookings
+  class CreateFirstCheckinMissionJob < ApplicationJob
+    queue_as :default
+
+    def perform(booking)
+      mission = Mission.create_first_checkin_for(booking)
+    end
+  end
+end
