+++ app/jobs/bookings/create_last_checkout_mission_job.rb
@@ -1,0 +1,9 @@
+module Bookings
+  class CreateLastCheckoutMissionJob < ApplicationJob
+    queue_as :default
+
+    def perform(booking)
+      Mission.create_last_checkout_for(booking)
+    end
+  end
+end
