--- app/jobs/create_first_checkin_mission_job.rb
@@ -1,7 +1,0 @@
-class CreateFirstCheckinMissionJob < ApplicationJob
-  queue_as :default
-
-  def perform(booking)
-    mission = Mission.create_first_checkin_for(booking)
-  end
-end
