--- app/jobs/create_last_checkout_mission_job.rb
@@ -1,7 +1,0 @@
-class CreateLastCheckoutMissionJob < ApplicationJob
-  queue_as :default
-
-  def perform(booking)
-    Mission.create_last_checkout_for(booking)
-  end
-end
