+++ app/jobs/handle_created_booking_job.rb
@@ -1,0 +1,8 @@
+class HandleCreatedBookingJob < ApplicationJob
+  queue_as :default
+
+  def perform(booking)
+    Bookings::CreateFirstCheckinMissionJob.perform_now(booking)
+    Bookings::CreateLastCheckoutMissionJob.perform_now(booking)
+  end
+end
