+++ app/models/booking.rb
@@ -3,9 +3,7 @@ class Booking < ApplicationRecord
 
   belongs_to :listing
 
-  after_create -> { CreateFirstCheckinMissionJob.perform_now(self) }
-  after_create -> { CreateLastCheckoutMissionJob.perform_now(self) }
-
+  after_create -> { HandleCreatedBookingJob.perform_now(self) }
   after_destroy -> { HandleDestroyedBookingJob.perform_now(self) }
 
   def self.import(booking_attributes)
