+++ test/jobs/bookings/create_first_checkin_mission_job_test.rb
@@ -1,19 +1,19 @@
 require "test_helper"
 
-class CreateFirstCheckinMissionJobTest < ActiveJob::TestCase
+class Bookings::CreateFirstCheckinMissionJobTest < ActiveJob::TestCase
   setup do
     @booking = bookings(:two)
   end
 
   test "create a mission" do
     assert_difference 'Mission.count' do
-      CreateFirstCheckinMissionJob.perform_now(@booking)
+      Bookings::CreateFirstCheckinMissionJob.perform_now(@booking)
     end
 
   end
 
   test "sets the mission attributes" do
-    mission = CreateFirstCheckinMissionJob.perform_now(@booking)
+    mission = Bookings::CreateFirstCheckinMissionJob.perform_now(@booking)
 
     assert_predicate mission, :first_checkin?
     assert_equal mission.date, @booking.start_date
