+++ test/jobs/handle_created_booking_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class HandleCreatedBookingJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
